Add files via upload
Take out the data from the quicker of some car its arounding 900 car data are available on the basis of only this much of the data we prepare a model that gives a prediction value of the price. Because the data set is very low then prediction accuracy is not much well as we want. Its only 72% predicton accuracy which is not much good.